requirements.txt
streamlit
pandas
psycopg2-binary
matplotlib
matplotlib-venn
plotly
openpyxl
xlsxwriter